cadastro e consulta de genres e movies com relacionamento entre ele

+++ src/main/java/br/com/meli/movies1/repository/IGenresRepository.java
@@ -4,5 +4,5 @@ package br.com.meli.movies1.repository;
 import br.com.meli.movies1.model.Genres;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IGenresRepository extends JpaRepository<Genres, Long> {
+public interface IGenresRepository extends JpaRepository<Genres, Integer> {
 }
